Move stats test into separate file
